requirements.txt
flask==3.0.0
python-dotenv==1.0.0
openai==1.3.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
